First Commit
Created feature vector that contains each unique data dimension (total # of dimensions is 258).  I also noticed that np.genfromtxt only generates a one-dimensional array where each line in the txt file is treated as one giant element.  I experimented a lot with this and found that np.column_stack is able to break this one-dimensional array into a two-dimensional array where each characteristic (age, education level, etc.) is indexed in a unique row and column.  If there is a better way to do this, feel to modify/change completely.  -Peter