Improved Binarized Data/Basic Perceptron
Figured out how to converted data file into 2-dimensional array.

Created the basic perceptron algorithm (haven't played with the dev data, all it does is iterates through training data for a set number of epochs).   I realized halfway through that cross-contamination was occurring between the age and work hours features (since, for example, each of those categories contained the number 50), so I had to add the strings 'Age' and 'Hours' to those features.  I'm not sure if that is the best way to fix this problem.

Anyways, in order to save memory, I only initialize the weight vector with the full 259 features.  Whenever the weight vector is updated, I indexed the weight vector entries that corresponded to the training datum's features and just summed those weight vector entries.  I did this because all of the features are binary currently, so the dot product of the weight vector with the training data is just going to be the sum of  weight vector elements corresponding to the nine features the training datum possesses.

Again, feel free to make any changes, I'm really new to Python so there are probably more elegant ways to implement this algorithm.  -Peter